mmu nearly running

+++ kernel/src/task/taskTable.c
@@ -9,10 +9,10 @@
 
 static Task taskTable[ MAX_TASKS ];
 
-static uint32_t numTasks = 0;
-static uint32_t nextFreePID = 0;
+static uint8_t numTasks = 0;
+static uint8_t nextFreePID = 0;
 
-Task* getTask( uint32_t pid )
+Task* getTask( uint8_t pid )
 {
 	if( pid > MAX_TASKS ) {
 		return ( void* ) 0;
@@ -45,13 +45,10 @@ int32_t addTask( Task* task )
 	return 0;
 }
 
-int32_t getNextFreePID() {
+uint8_t getNextFreePID() {
 	return nextFreePID;
 }
 
-int32_t getNumOfTasks() {
+uint8_t getNumOfTasks() {
 	return numTasks;
 }
-
-
-
